refactor file_io.c code
