+++ src/communication/network_interface_cl.c
@@ -67,6 +67,7 @@
 #include "db_query.h"
 #include "dbtype.h"
 #include "compile_context.h"
+#include "file_io.h"
 
 #if defined (SA_MODE)
 #include "thread_manager.hpp"
