+++ src/communication/network_interface_sr.c
@@ -11556,4 +11556,3 @@ smanage_lob_dir (THREAD_ENTRY * thread_p, unsigned int rid, char *request, int r
   ptr = or_pack_errcode (reply, error);
   css_send_data_to_client (thread_p->conn_entry, rid, reply, OR_ALIGNED_BUF_SIZE (a_reply));
 }
-
