+++ src/storage/file_io.h
@@ -37,6 +37,7 @@
 #include "release_string.h"
 #include "storage_common.h"
 #include "thread_compat.hpp"
+#include "recovery.h"
 
 #include <stdio.h>
 #include <time.h>
@@ -631,4 +632,11 @@ extern int fileio_set_page_checksum (THREAD_ENTRY * thread_p, FILEIO_PAGE * io_p
 extern int fileio_page_check_corruption (THREAD_ENTRY * thread_p, FILEIO_PAGE * io_page, bool * is_page_corrupted);
 extern void fileio_page_hexa_dump (const char *data, int length);
 extern bool fileio_is_formatted_page (THREAD_ENTRY * thread_p, const char *io_page);
+
+// #if defined(SERVER_MODE)
+/* lob_dir */
+extern int fileio_lob_dir_remove (const char *path);
+extern void xmanage_lob_dir (HFID * hfid, int * attrid_arr, int lob_arr_length, LOB_DIR_MANAGE_MODE mode);
+extern int fileio_lob_rv_destroy (THREAD_ENTRY * thread_p, LOG_RCV * rcv);
+// #endif /* SERVER_MODE */
 #endif /* _FILE_IO_H_ */
